requirements.txt
python-telegram-bot>=22.5
requests>=2.32.5
APScheduler>=3.11.0
duckduckgo-search>=8.1.1
bing-image-downloader>=1.1.2
python-dotenv
